+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changes between versions
 
-## Not yet released
+## 1.2.0 (2017-07-17)
 
 * Added support for sound and subtitle option on AppleScript notifier
 * Added support for url option on TerminalNotifier notifier